add rnn test
